tools/pysflib/requirements.txt
nose
mock
requests>=2.4.1
python-redmine
pyyaml
six
-e git+https://github.com/morucci/pygerrit.git#egg=pygerrit
